mysite/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.11.0
gunicorn==20.0.4
PyMySQL==0.9.3
pytz==2019.3
screen==1.0.1
sqlparse==0.3.1
uWSGI==2.0.18
xlwt==1.3.0
